refactor: 💡 update Entity files

See merge request s07-bigdata-recom-sub2/S07P22B206!33

+++ yogidice/src/main/java/com/specialization/yogidice/dto/request/category/MechanismGroupCreateRequest.java
@@ -14,5 +14,5 @@ public class MechanismGroupCreateRequest {
     private Long gameId;
 
     @NotNull
-    private Long mechanismId;
+    private Byte mechanismId;
 }
